refactor: move SuggestedChartRevisionStatus enum to clientUtils
