+++ adminSiteClient/SuggestedChartRevision.ts
@@ -1,11 +1,5 @@
 import { GrapherInterface } from "../grapher/core/GrapherInterface"
-
-export enum SuggestedChartRevisionStatus {
-    pending = "pending",
-    approved = "approved",
-    rejected = "rejected",
-    flagged = "flagged",
-}
+import { SuggestedChartRevisionStatus } from "../clientUtils/owidTypes"
 
 export interface SuggestedChartRevisionSerialized {
     id: number
