+++ clientUtils/owidTypes.ts
@@ -265,6 +265,13 @@ export interface FullPost {
     glossary: boolean
 }
 
+export enum SuggestedChartRevisionStatus {
+    pending = "pending",
+    approved = "approved",
+    rejected = "rejected",
+    flagged = "flagged",
+}
+
 // Exception format that can be easily given as an API error
 export class JsonError extends Error {
     status: number
